backend/requirements.txt
Flask==2.3.3
Flask-PyMongo==2.3.0
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
pymongo==4.5.0
boto3==1.29.7
prometheus-client==0.17.1
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
